pyproject.toml
[tool.poetry]
name = "receipts"
version = "0.1.0"
description = ""
authors = ["gary_meehan <garymeehan@outlook.com>", "sam_meehan <sammeehansam@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = {extras = ["standard"], version = "^0.115.6"}
uvicorn = {extras = ["standard"], version = "^0.34.0"}
google-cloud-documentai = "^3.1.0"
python-multipart = "^0.0.20"
python-dotenv = "^1.0.1"
pydantic-settings = "^2.7.0"


[tool.poetry.group.dev.dependencies]
mypy = "^1.13.0"
pydantic = {extras = ["mypy"], version = "^2.10.3"}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
